motor

+++ projects/DHT11AndDoorWithRF433/DHT11AndDoorWithRF433.ino
@@ -91,9 +91,9 @@ void loop() {
   if (sleepCounter == 0) {
     // Reading temperature or humidity takes about 250 milliseconds!
     // Sensor readings may also be up to 2 seconds 'old' (its a very slow sensor)
-    int h = dht.readHumidity();
+    float h = dht.readHumidity();
     // Read temperature as Celsius (the default)
-    int t = dht.readTemperature();
+    float t = dht.readTemperature();
   
     // Check if any reads failed and exit early (to try again).
     if (isnan(h) || isnan(t)) {
@@ -104,7 +104,7 @@ void loop() {
     int v = readVcc();
   
     // Compute heat index in Celsius (isFahreheit = false)
-    int hic = dht.computeHeatIndex(t, h, false);
+    float hic = dht.computeHeatIndex(t, h, false);
 
     // Get Door Sensor Status
     int doorStatus = digitalRead(doorPin);
